This is release of the devkin application

+++ devkin/src/main/java/com/example/devkin/services/FileStorageService.java
@@ -1,5 +1,6 @@
 package com.example.devkin.services;
 
+import com.example.devkin.dtos.FileDto;
 import com.example.devkin.entities.File;
 import com.example.devkin.entities.Folder;
 import com.example.devkin.entities.Project;
@@ -7,18 +8,16 @@ import com.example.devkin.repositories.FileRepository;
 import com.example.devkin.repositories.FolderRepository;
 import com.example.devkin.repositories.ProjectRepository;
 import io.minio.*;
-import io.minio.errors.MinioException;
 import io.minio.messages.Item;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
-
 import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
 import java.time.LocalDateTime;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
 
 @Service
 public class FileStorageService {
@@ -59,6 +58,7 @@ public class FileStorageService {
             folder.setProject(projectRepository.findById(projectId).orElseThrow(() -> new RuntimeException("Project not found")));
 
             folderRepository.save(folder);
+
         } catch (Exception e) {
             throw new RuntimeException("Failed to create folder", e);
         }
@@ -146,6 +146,11 @@ public class FileStorageService {
         try {
             String filePath = folderPath.endsWith("/") ? folderPath + fileName : folderPath + "/" + fileName;
 
+            boolean fileExists = fileRepository.existsByFilePath(filePath);
+            if (fileExists) {
+                throw new RuntimeException("File already exists at this path: " + filePath);
+            }
+
             minioClient.putObject(
                     PutObjectArgs.builder()
                             .bucket("devkin")
@@ -168,12 +173,15 @@ public class FileStorageService {
                     .orElseThrow(() -> new RuntimeException("Project not found")));
 
             fileRepository.save(file);
+
+            // Update project size
             Optional<Project> p = projectRepository.findById(projectId);
             p.get().setSize(p.get().getSize() + fileSize);
         } catch (Exception e) {
             throw new RuntimeException("Failed to create file", e);
         }
     }
+
 
     @Transactional
     public void updateFileName(String oldFilePath, String newFileName) {
@@ -240,5 +248,50 @@ public class FileStorageService {
         } catch (Exception e) {
             throw new RuntimeException("Failed to delete file", e);
         }
+    }
+
+    @Transactional
+    public byte[] getFileContents(String filePath) {
+        try {
+            // Download the file as an InputStream from Minio
+            GetObjectArgs getObjectArgs = GetObjectArgs.builder()
+                    .bucket("devkin")
+                    .object(filePath)
+                    .build();
+            try (var inputStream = minioClient.getObject(getObjectArgs)) {
+                return inputStream.readAllBytes();
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("Failed to get file contents", e);
+        }
+    }
+
+    @Transactional
+    public String getFileContentType(String filePath) {
+        try {
+            StatObjectResponse stat = minioClient.statObject(StatObjectArgs.builder()
+                    .bucket("devkin")
+                    .object(filePath)
+                    .build());
+
+            return stat.contentType();
+        } catch (Exception e) {
+            throw new RuntimeException("Failed to get file content type", e);
+        }
+    }
+
+    public String calcFilePath(FileDto fileDto){
+        try {
+            Project project = projectRepository.findBySlug(fileDto.getProjectSlug()).orElseThrow(() -> new RuntimeException("Project not found"));
+            Integer ownerId = project.getOwner().getId();
+
+            return "projects/" + ownerId + "/" + project.getName() +
+                    (fileDto.getFilePath().isEmpty() ? "" : "/" + fileDto.getFilePath()) +
+                    "/" + fileDto.getFileName();
+        }
+        catch (Exception e){
+            throw new RuntimeException("Can't calculate the file path");
+        }
+
     }
 }
